+++ rust/operator-binary/src/discovery.rs
@@ -1,18 +1,43 @@
-use crate::{
-    build_recommended_labels,
-    config::{CoreSiteConfigBuilder, HdfsSiteConfigBuilder},
-    hdfs_controller::Error,
-};
+use snafu::{ResultExt, Snafu};
 use stackable_hdfs_crd::{
     constants::{CORE_SITE_XML, HDFS_SITE_XML},
     HdfsCluster, HdfsPodRef, HdfsRole,
 };
 use stackable_operator::{
-    builder::{ConfigMapBuilder, ObjectMetaBuilder},
+    builder::{ConfigMapBuilder, ObjectMetaBuilder, ObjectMetaBuilderError},
     commons::product_image_selection::ResolvedProductImage,
     k8s_openapi::api::core::v1::ConfigMap,
     kube::{runtime::reflector::ObjectRef, ResourceExt},
 };
+
+use crate::{
+    build_recommended_labels,
+    config::{CoreSiteConfigBuilder, HdfsSiteConfigBuilder},
+    kerberos,
+};
+
+type Result<T, E = Error> = std::result::Result<T, E>;
+
+#[derive(Snafu, Debug)]
+#[allow(clippy::enum_variant_names)]
+pub enum Error {
+    #[snafu(display("object {hdfs} is missing metadata to build owner reference"))]
+    ObjectMissingMetadataForOwnerRef {
+        source: stackable_operator::error::Error,
+        hdfs: ObjectRef<HdfsCluster>,
+    },
+
+    #[snafu(display("failed to build ConfigMap"))]
+    BuildConfigMap {
+        source: stackable_operator::error::Error,
+    },
+
+    #[snafu(display("failed to build object meta data"))]
+    ObjectMeta { source: ObjectMetaBuilderError },
+
+    #[snafu(display("failed to build security discovery config map"))]
+    BuildSecurityDiscoveryConfigMap { source: kerberos::Error },
+}
 
 /// Creates a discovery config map containing the `hdfs-site.xml` and `core-site.xml`
 /// for clients.
@@ -21,25 +46,25 @@ pub fn build_discovery_configmap(
     controller: &str,
     namenode_podrefs: &[HdfsPodRef],
     resolved_product_image: &ResolvedProductImage,
-) -> Result<ConfigMap, crate::hdfs_controller::Error> {
+) -> Result<ConfigMap> {
+    let metadata = ObjectMetaBuilder::new()
+        .name_and_namespace(hdfs)
+        .ownerreference_from_resource(hdfs, None, Some(true))
+        .context(ObjectMissingMetadataForOwnerRefSnafu {
+            hdfs: ObjectRef::from_obj(hdfs),
+        })?
+        .with_recommended_labels(build_recommended_labels(
+            hdfs,
+            controller,
+            &resolved_product_image.app_version_label,
+            &HdfsRole::NameNode.to_string(),
+            "discovery",
+        ))
+        .context(ObjectMetaSnafu)?
+        .build();
+
     ConfigMapBuilder::new()
-        .metadata(
-            ObjectMetaBuilder::new()
-                .name_and_namespace(hdfs)
-                .ownerreference_from_resource(hdfs, None, Some(true))
-                .map_err(|err| Error::ObjectMissingMetadataForOwnerRef {
-                    source: err,
-                    obj_ref: ObjectRef::from_obj(hdfs),
-                })?
-                .with_recommended_labels(build_recommended_labels(
-                    hdfs,
-                    controller,
-                    &resolved_product_image.app_version_label,
-                    &HdfsRole::NameNode.to_string(),
-                    "discovery",
-                ))
-                .build(),
-        )
+        .metadata(metadata)
         .add_data(
             HDFS_SITE_XML,
             build_discovery_hdfs_site_xml(hdfs, hdfs.name_any(), namenode_podrefs),
@@ -49,7 +74,7 @@ pub fn build_discovery_configmap(
             build_discovery_core_site_xml(hdfs, hdfs.name_any())?,
         )
         .build()
-        .map_err(|err| Error::BuildDiscoveryConfigMap { source: err })
+        .context(BuildConfigMapSnafu)
 }
 
 fn build_discovery_hdfs_site_xml(
@@ -67,12 +92,10 @@ fn build_discovery_hdfs_site_xml(
         .build_as_xml()
 }
 
-fn build_discovery_core_site_xml(
-    hdfs: &HdfsCluster,
-    logical_name: String,
-) -> Result<String, Error> {
+fn build_discovery_core_site_xml(hdfs: &HdfsCluster, logical_name: String) -> Result<String> {
     Ok(CoreSiteConfigBuilder::new(logical_name)
         .fs_default_fs()
-        .security_discovery_config(hdfs)?
+        .security_discovery_config(hdfs)
+        .context(BuildSecurityDiscoveryConfigMapSnafu)?
         .build_as_xml())
 }