+++ rust/operator-binary/src/hdfs_controller.rs
@@ -14,7 +14,10 @@ use stackable_hdfs_crd::{
     constants::*, AnyNodeConfig, HdfsCluster, HdfsClusterStatus, HdfsPodRef, HdfsRole,
 };
 use stackable_operator::{
-    builder::{ConfigMapBuilder, ObjectMetaBuilder, PodBuilder, PodSecurityContextBuilder},
+    builder::{
+        ConfigMapBuilder, ObjectMetaBuilder, ObjectMetaBuilderError, PodBuilder,
+        PodSecurityContextBuilder,
+    },
     client::Client,
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
     commons::{
@@ -34,7 +37,7 @@ use stackable_operator::{
         runtime::{controller::Action, reflector::ObjectRef},
         Resource, ResourceExt,
     },
-    labels::role_group_selector_labels,
+    kvp::{Label, LabelError, Labels},
     logging::controller::ReconcilerError,
     product_config_utils::{transform_all_roles_to_config, validate_all_roles_and_groups_config},
     role_utils::{GenericRoleConfig, RoleGroupRef},
@@ -49,9 +52,9 @@ use strum::{EnumDiscriminants, IntoStaticStr};
 use crate::{
     build_recommended_labels,
     config::{CoreSiteConfigBuilder, HdfsSiteConfigBuilder},
-    container::ContainerConfig,
+    container::{self, ContainerConfig},
     container::{TLS_STORE_DIR, TLS_STORE_PASSWORD},
-    discovery::build_discovery_configmap,
+    discovery::{self, build_discovery_configmap},
     event::{build_invalid_replica_message, publish_event},
     kerberos,
     operations::{
@@ -69,97 +72,92 @@ const DOCKER_IMAGE_BASE_NAME: &str = "hadoop";
 #[derive(Snafu, Debug, EnumDiscriminants)]
 #[strum_discriminants(derive(IntoStaticStr))]
 pub enum Error {
-    #[snafu(display("Invalid role configuration"))]
+    #[snafu(display("invalid role configuration"))]
     InvalidRoleConfig {
         source: stackable_operator::product_config_utils::ConfigError,
     },
 
-    #[snafu(display("Invalid product configuration"))]
+    #[snafu(display("invalid product configuration"))]
     InvalidProductConfig {
         source: stackable_operator::error::Error,
     },
 
-    #[snafu(display("Cannot create rolegroup service [{name}]"))]
+    #[snafu(display("cannot create rolegroup service {name:?}"))]
     ApplyRoleGroupService {
         source: stackable_operator::error::Error,
         name: String,
     },
 
-    #[snafu(display("Cannot create role group config map [{name}]"))]
+    #[snafu(display("cannot create role group config map {name:?}"))]
     ApplyRoleGroupConfigMap {
         source: stackable_operator::error::Error,
         name: String,
     },
 
-    #[snafu(display("Cannot create role group stateful set [{name}]"))]
+    #[snafu(display("cannot create role group stateful set {name:?}"))]
     ApplyRoleGroupStatefulSet {
         source: stackable_operator::error::Error,
         name: String,
     },
 
-    #[snafu(display("Cannot create discovery config map [{name}]"))]
+    #[snafu(display("cannot create discovery config map {name:?}"))]
     ApplyDiscoveryConfigMap {
         source: stackable_operator::error::Error,
         name: String,
     },
 
-    #[snafu(display("No metadata for [{obj_ref}]"))]
+    #[snafu(display("no metadata for {obj_ref:?}"))]
     ObjectMissingMetadataForOwnerRef {
         source: stackable_operator::error::Error,
         obj_ref: ObjectRef<HdfsCluster>,
     },
 
-    #[snafu(display("Invalid role [{role}]"))]
+    #[snafu(display("invalid role {role:?}"))]
     InvalidRole {
         source: strum::ParseError,
         role: String,
     },
 
-    #[snafu(display("Object has no name"))]
+    #[snafu(display("object has no name"))]
     ObjectHasNoName { obj_ref: ObjectRef<HdfsCluster> },
 
-    #[snafu(display("Object has no namespace"))]
-    ObjectHasNoNamespace { obj_ref: ObjectRef<HdfsCluster> },
-
-    #[snafu(display("Cannot build config map for role [{role}] and role group [{role_group}]"))]
+    #[snafu(display("cannot build config map for role {role:?} and role group {role_group:?}"))]
     BuildRoleGroupConfigMap {
         source: stackable_operator::error::Error,
         role: String,
         role_group: String,
     },
 
-    #[snafu(display("Cannot collect discovery configuration"))]
+    #[snafu(display("cannot collect discovery configuration"))]
     CollectDiscoveryConfig { source: stackable_hdfs_crd::Error },
 
-    #[snafu(display("Cannot build discovery config map"))]
-    BuildDiscoveryConfigMap {
-        source: stackable_operator::error::Error,
-    },
+    #[snafu(display("cannot build config discovery config map"))]
+    BuildDiscoveryConfigMap { source: discovery::Error },
 
-    #[snafu(display("Failed to patch service account"))]
+    #[snafu(display("failed to patch service account"))]
     ApplyServiceAccount {
         source: stackable_operator::error::Error,
     },
 
-    #[snafu(display("Failed to patch role binding"))]
+    #[snafu(display("failed to patch role binding"))]
     ApplyRoleBinding {
         source: stackable_operator::error::Error,
     },
 
-    #[snafu(display("Failed to create cluster resources"))]
+    #[snafu(display("failed to create cluster resources"))]
     CreateClusterResources {
         source: stackable_operator::error::Error,
     },
 
-    #[snafu(display("Failed to delete orphaned resources"))]
+    #[snafu(display("failed to delete orphaned resources"))]
     DeleteOrphanedResources {
         source: stackable_operator::error::Error,
     },
 
-    #[snafu(display("Failed to create pod references"))]
+    #[snafu(display("failed to create pod references"))]
     CreatePodReferences { source: stackable_hdfs_crd::Error },
 
-    #[snafu(display("Failed to build role properties"))]
+    #[snafu(display("failed to build role properties"))]
     BuildRoleProperties { source: stackable_hdfs_crd::Error },
 
     #[snafu(display("failed to resolve the Vector aggregator address"))]
@@ -167,7 +165,7 @@ pub enum Error {
         source: crate::product_logging::Error,
     },
 
-    #[snafu(display("failed to add the logging configuration to the ConfigMap [{cm_name}]"))]
+    #[snafu(display("failed to add the logging configuration to the ConfigMap {cm_name:?}"))]
     InvalidLoggingConfig {
         source: crate::product_logging::Error,
         cm_name: String,
@@ -203,7 +201,7 @@ pub enum Error {
     KerberosNotSupported,
 
     #[snafu(display(
-        "failed to serialize [{JVM_SECURITY_PROPERTIES_FILE}] for {}",
+        "failed to serialize {JVM_SECURITY_PROPERTIES_FILE:?} for {}",
         rolegroup
     ))]
     JvmSecurityProperties {
@@ -213,6 +211,27 @@ pub enum Error {
 
     #[snafu(display("failed to configure graceful shutdown"))]
     GracefulShutdown { source: graceful_shutdown::Error },
+
+    #[snafu(display("failed to build roleGroup selector labels"))]
+    RoleGroupSelectorLabels { source: stackable_hdfs_crd::Error },
+
+    #[snafu(display("failed to build prometheus label"))]
+    BuildPrometheusLabel { source: LabelError },
+
+    #[snafu(display("failed to build cluster resources label"))]
+    BuildClusterResourcesLabel { source: LabelError },
+
+    #[snafu(display("failed to build role-group selector label"))]
+    BuildRoleGroupSelectorLabel { source: LabelError },
+
+    #[snafu(display("failed to build role-group volume claim templates from config"))]
+    BuildRoleGroupVolumeClaimTemplates { source: container::Error },
+
+    #[snafu(display("failed to build object meta data"))]
+    ObjectMeta { source: ObjectMetaBuilderError },
+
+    #[snafu(display("failed to build security config"))]
+    BuildSecurityConfig { source: kerberos::Error },
 }
 
 impl ReconcilerError for Error {
@@ -236,8 +255,8 @@ pub async fn reconcile_hdfs(hdfs: Arc<HdfsCluster>, ctx: Arc<Ctx>) -> HdfsOperat
         .spec
         .image
         .resolve(DOCKER_IMAGE_BASE_NAME, crate::built_info::CARGO_PKG_VERSION);
-    if hdfs.has_kerberos_enabled() {
-        kerberos::check_if_supported(&resolved_product_image)?;
+    if hdfs.has_kerberos_enabled() && kerberos::is_not_supported(&resolved_product_image) {
+        return KerberosNotSupportedSnafu.fail();
     }
 
     let vector_aggregator_address = resolve_vector_aggregator_address(&hdfs, client)
@@ -279,7 +298,9 @@ pub async fn reconcile_hdfs(hdfs: Arc<HdfsCluster>, ctx: Arc<Ctx>) -> HdfsOperat
     let (rbac_sa, rbac_rolebinding) = build_rbac_resources(
         hdfs.as_ref(),
         APP_NAME,
-        cluster_resources.get_required_labels(),
+        cluster_resources
+            .get_required_labels()
+            .context(BuildClusterResourcesLabelSnafu)?,
     )
     .context(BuildRbacResourcesSnafu)?;
 
@@ -390,7 +411,8 @@ pub async fn reconcile_hdfs(hdfs: Arc<HdfsCluster>, ctx: Arc<Ctx>) -> HdfsOperat
             .await
             .context(CollectDiscoveryConfigSnafu)?,
         &resolved_product_image,
-    )?;
+    )
+    .context(BuildDiscoveryConfigMapSnafu)?;
 
     // The discovery CM is linked to the cluster lifecycle via ownerreference.
     // Therefore, must not be added to the "orphaned" cluster resources
@@ -430,42 +452,55 @@ fn rolegroup_service(
     resolved_product_image: &ResolvedProductImage,
 ) -> HdfsOperatorResult<Service> {
     tracing::info!("Setting up Service for {:?}", rolegroup_ref);
+
+    let prometheus_label =
+        Label::try_from(("prometheus.io/scrape", "true")).context(BuildPrometheusLabelSnafu)?;
+
+    let metadata = ObjectMetaBuilder::new()
+        .name_and_namespace(hdfs)
+        .name(&rolegroup_ref.object_name())
+        .ownerreference_from_resource(hdfs, None, Some(true))
+        .context(ObjectMissingMetadataForOwnerRefSnafu {
+            obj_ref: ObjectRef::from_obj(hdfs),
+        })?
+        .with_recommended_labels(build_recommended_labels(
+            hdfs,
+            RESOURCE_MANAGER_HDFS_CONTROLLER,
+            &resolved_product_image.app_version_label,
+            &rolegroup_ref.role,
+            &rolegroup_ref.role_group,
+        ))
+        .context(ObjectMetaSnafu)?
+        .with_label(prometheus_label)
+        .build();
+
+    let service_spec = ServiceSpec {
+        // Internal communication does not need to be exposed
+        type_: Some("ClusterIP".to_string()),
+        cluster_ip: Some("None".to_string()),
+        ports: Some(
+            hdfs.ports(role)
+                .into_iter()
+                .map(|(name, value)| ServicePort {
+                    name: Some(name),
+                    port: i32::from(value),
+                    protocol: Some("TCP".to_string()),
+                    ..ServicePort::default()
+                })
+                .collect(),
+        ),
+        selector: Some(
+            hdfs.rolegroup_selector_labels(rolegroup_ref)
+                .context(RoleGroupSelectorLabelsSnafu)?
+                .into(),
+        ),
+        publish_not_ready_addresses: Some(true),
+        ..ServiceSpec::default()
+    };
+
     Ok(Service {
-        metadata: ObjectMetaBuilder::new()
-            .name_and_namespace(hdfs)
-            .name(&rolegroup_ref.object_name())
-            .ownerreference_from_resource(hdfs, None, Some(true))
-            .with_context(|_| ObjectMissingMetadataForOwnerRefSnafu {
-                obj_ref: ObjectRef::from_obj(hdfs),
-            })?
-            .with_recommended_labels(build_recommended_labels(
-                hdfs,
-                RESOURCE_MANAGER_HDFS_CONTROLLER,
-                &resolved_product_image.app_version_label,
-                &rolegroup_ref.role,
-                &rolegroup_ref.role_group,
-            ))
-            .with_label("prometheus.io/scrape", "true")
-            .build(),
-        spec: Some(ServiceSpec {
-            // Internal communication does not need to be exposed
-            type_: Some("ClusterIP".to_string()),
-            cluster_ip: Some("None".to_string()),
-            ports: Some(
-                hdfs.ports(role)
-                    .into_iter()
-                    .map(|(name, value)| ServicePort {
-                        name: Some(name),
-                        port: i32::from(value),
-                        protocol: Some("TCP".to_string()),
-                        ..ServicePort::default()
-                    })
-                    .collect(),
-            ),
-            selector: Some(hdfs.rolegroup_selector_labels(rolegroup_ref)),
-            publish_not_ready_addresses: Some(true),
-            ..ServiceSpec::default()
-        }),
+        metadata,
+        spec: Some(service_spec),
         status: None,
     })
 }
@@ -551,7 +586,8 @@ fn rolegroup_config_map(
                 core_site_xml = CoreSiteConfigBuilder::new(hdfs_name.to_string())
                     .fs_default_fs()
                     .ha_zookeeper_quorum()
-                    .security_config(hdfs)?
+                    .security_config(hdfs)
+                    .context(BuildSecurityConfigSnafu)?
                     // the extend with config must come last in order to have overrides working!!!
                     .extend(config)
                     .build_as_xml();
@@ -632,24 +668,25 @@ fn rolegroup_config_map(
         .map(|(k, v)| (k, Some(v)))
         .collect();
 
+    let cm_metadata = ObjectMetaBuilder::new()
+        .name_and_namespace(hdfs)
+        .name(&rolegroup_ref.object_name())
+        .ownerreference_from_resource(hdfs, None, Some(true))
+        .with_context(|_| ObjectMissingMetadataForOwnerRefSnafu {
+            obj_ref: ObjectRef::from_obj(hdfs),
+        })?
+        .with_recommended_labels(build_recommended_labels(
+            hdfs,
+            RESOURCE_MANAGER_HDFS_CONTROLLER,
+            &resolved_product_image.app_version_label,
+            &rolegroup_ref.role,
+            &rolegroup_ref.role_group,
+        ))
+        .context(ObjectMetaSnafu)?
+        .build();
+
     builder
-        .metadata(
-            ObjectMetaBuilder::new()
-                .name_and_namespace(hdfs)
-                .name(&rolegroup_ref.object_name())
-                .ownerreference_from_resource(hdfs, None, Some(true))
-                .with_context(|_| ObjectMissingMetadataForOwnerRefSnafu {
-                    obj_ref: ObjectRef::from_obj(hdfs),
-                })?
-                .with_recommended_labels(build_recommended_labels(
-                    hdfs,
-                    RESOURCE_MANAGER_HDFS_CONTROLLER,
-                    &resolved_product_image.app_version_label,
-                    &rolegroup_ref.role,
-                    &rolegroup_ref.role_group,
-                ))
-                .build(),
-        )
+        .metadata(cm_metadata)
         .add_data(CORE_SITE_XML.to_string(), core_site_xml)
         .add_data(HDFS_SITE_XML.to_string(), hdfs_site_xml)
         .add_data(HADOOP_POLICY_XML.to_string(), hadoop_policy_xml)
@@ -697,20 +734,27 @@ fn rolegroup_statefulset(
     let object_name = rolegroup_ref.object_name();
     // PodBuilder for StatefulSet Pod template.
     let mut pb = PodBuilder::new();
-    pb.metadata(ObjectMeta {
-        labels: Some(hdfs.rolegroup_selector_labels(rolegroup_ref)),
+
+    let pb_metadata = ObjectMeta {
+        labels: Some(
+            hdfs.rolegroup_selector_labels(rolegroup_ref)
+                .context(RoleGroupSelectorLabelsSnafu)?
+                .into(),
+        ),
         ..ObjectMeta::default()
-    })
-    .image_pull_secrets_from_product_image(resolved_product_image)
-    .affinity(&merged_config.affinity)
-    .service_account_name(service_account_name(APP_NAME))
-    .security_context(
-        PodSecurityContextBuilder::new()
-            .run_as_user(HDFS_UID)
-            .run_as_group(0)
-            .fs_group(1000)
-            .build(),
-    );
+    };
+
+    pb.metadata(pb_metadata)
+        .image_pull_secrets_from_product_image(resolved_product_image)
+        .affinity(&merged_config.affinity)
+        .service_account_name(service_account_name(APP_NAME))
+        .security_context(
+            PodSecurityContextBuilder::new()
+                .run_as_user(HDFS_UID)
+                .run_as_group(0)
+                .fs_group(1000)
+                .build(),
+        );
 
     // Adds all containers and volumes to the pod builder
     ContainerConfig::add_containers_and_volumes(
@@ -737,42 +781,53 @@ fn rolegroup_statefulset(
         pod_template.merge_from(pod_overrides.clone());
     }
 
-    Ok(StatefulSet {
-        metadata: ObjectMetaBuilder::new()
-            .name_and_namespace(hdfs)
-            .name(&rolegroup_ref.object_name())
-            .ownerreference_from_resource(hdfs, None, Some(true))
-            .with_context(|_| ObjectMissingMetadataForOwnerRefSnafu {
-                obj_ref: ObjectRef::from_obj(hdfs),
-            })?
-            .with_recommended_labels(build_recommended_labels(
-                hdfs,
-                RESOURCE_MANAGER_HDFS_CONTROLLER,
-                &resolved_product_image.app_version_label,
-                &rolegroup_ref.role,
-                &rolegroup_ref.role_group,
-            ))
-            .build(),
-        spec: Some(StatefulSetSpec {
-            pod_management_policy: Some("OrderedReady".to_string()),
-            replicas: role
-                .role_group_replicas(hdfs, &rolegroup_ref.role_group)
-                .map(i32::from),
-            selector: LabelSelector {
-                match_labels: Some(role_group_selector_labels(
-                    hdfs,
-                    APP_NAME,
-                    &rolegroup_ref.role,
-                    &rolegroup_ref.role_group,
-                )),
-                ..LabelSelector::default()
-            },
-            service_name: object_name,
-            template: pod_template,
+    let metadata = ObjectMetaBuilder::new()
+        .name_and_namespace(hdfs)
+        .name(&rolegroup_ref.object_name())
+        .ownerreference_from_resource(hdfs, None, Some(true))
+        .with_context(|_| ObjectMissingMetadataForOwnerRefSnafu {
+            obj_ref: ObjectRef::from_obj(hdfs),
+        })?
+        .with_recommended_labels(build_recommended_labels(
+            hdfs,
+            RESOURCE_MANAGER_HDFS_CONTROLLER,
+            &resolved_product_image.app_version_label,
+            &rolegroup_ref.role,
+            &rolegroup_ref.role_group,
+        ))
+        .context(ObjectMetaSnafu)?
+        .build();
 
-            volume_claim_templates: Some(ContainerConfig::volume_claim_templates(merged_config)),
-            ..StatefulSetSpec::default()
-        }),
+    let match_labels = Labels::role_group_selector(
+        hdfs,
+        APP_NAME,
+        &rolegroup_ref.role,
+        &rolegroup_ref.role_group,
+    )
+    .context(BuildRoleGroupSelectorLabelSnafu)?;
+
+    let pvcs = ContainerConfig::volume_claim_templates(merged_config)
+        .context(BuildRoleGroupVolumeClaimTemplatesSnafu)?;
+
+    let statefulset_spec = StatefulSetSpec {
+        pod_management_policy: Some("OrderedReady".to_string()),
+        replicas: role
+            .role_group_replicas(hdfs, &rolegroup_ref.role_group)
+            .map(i32::from),
+        selector: LabelSelector {
+            match_labels: Some(match_labels.into()),
+            ..LabelSelector::default()
+        },
+        service_name: object_name,
+        template: pod_template,
+
+        volume_claim_templates: Some(pvcs),
+        ..StatefulSetSpec::default()
+    };
+
+    Ok(StatefulSet {
+        metadata,
+        spec: Some(statefulset_spec),
         status: None,
     })
 }