+++ rust/operator-binary/src/operations/graceful_shutdown.rs
@@ -4,7 +4,7 @@ use stackable_operator::builder::PodBuilder;
 
 #[derive(Debug, Snafu)]
 pub enum Error {
-    #[snafu(display("Failed to set terminationGracePeriod"))]
+    #[snafu(display("failed to set terminationGracePeriod"))]
     SetTerminationGracePeriod {
         source: stackable_operator::builder::pod::Error,
     },