+++ rust/operator-binary/src/product_logging.rs
@@ -1,9 +1,7 @@
 use std::{borrow::Cow, fmt::Display};
 
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_hdfs_crd::{
-    AnyNodeConfig, DataNodeContainer, HdfsCluster, LoggingExt, NameNodeContainer,
-};
+use stackable_hdfs_crd::{AnyNodeConfig, DataNodeContainer, HdfsCluster, NameNodeContainer};
 use stackable_operator::{
     builder::ConfigMapBuilder,
     client::Client,
@@ -21,16 +19,19 @@ use stackable_operator::{
 pub enum Error {
     #[snafu(display("object has no namespace"))]
     ObjectHasNoNamespace,
-    #[snafu(display("failed to retrieve the ConfigMap [{cm_name}]"))]
+
+    #[snafu(display("failed to retrieve the ConfigMap {cm_name:?}"))]
     ConfigMapNotFound {
         source: stackable_operator::error::Error,
         cm_name: String,
     },
-    #[snafu(display("failed to retrieve the entry [{entry}] for ConfigMap [{cm_name}]"))]
+
+    #[snafu(display("failed to retrieve the entry {entry:?} for ConfigMap {cm_name:?}"))]
     MissingConfigMapEntry {
         entry: &'static str,
         cm_name: String,
     },
+
     #[snafu(display("vectorAggregatorConfigMapName must be set"))]
     MissingVectorAggregatorAddress,
 }